appengine/uploader/requirements.txt
absl-py==0.7.0
arrow==0.13.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
future==0.17.1
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-cloud-logging==1.10.0
google-cloud-pubsub==0.39.1
google-cloud-tasks==0.4.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==3.0.1
parameterized==0.7.0
pbr==5.1.1
pip==19.2.3
pipdeptree==0.13.2
ply==3.11
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.7.3
uritemplate==3.0.0
urllib3==1.24.1
waitress==1.2.1
WebOb==1.8.5
Werkzeug==0.14.1
